backend/requirements.txt
aiofiles==24.1.0
anyio==4.9.0
fastapi==0.116.1
google-genai==1.26.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
python-multipart
starlette==0.47.2
typing_extensions==4.14.1
uvicorn==0.35.0
